Judge: Abstract CMS judges
